requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
supabase==2.3.4
postgrest==0.16.0

